""9:20am. I am up. Let me chill for a while and then I will start.
https://www.reddit.com/r/ProgrammingLanguages/comments/f7cp5k/conor_mcbride_type_inference_needs_revolution/

Maybe I'll try watching this as well.

9:30am. https://mangadex.org/title/26399/tobaku-datenroku-kaiji-24-oku-dasshutsu-hen

Oh, lol. The stuff in the latest chapters is way beyond what I expected. God tier writing as usual. Every arc in Kaiji is the best arc.

10:45am. Ah, I started arguing on /sci/ and wasted my time. Yes, non-constructive math put deep scars on my soul and now I cannot help myself but to get in random math foundation fights. I'll try to do better next time.

But I might as well take a real break here since it has come to that.

12:20pm. 10m more and then I'll do the chores. Then I'll resume.

Actually, let me give the talk by Conor a try. I'll leave the Toji game for later.

1:10pm. Done with chores. Let me finally start this shit. I do not really like the talk, between the poor audio quality and the awkward delivery, I have better things to do. Maybe I'll watch it at some later point if I feel like it.

Now...I have a couple of good hours ahead of me. Let me focus on the book. Here things finally start getting real.

1:20pm. 106/558. The first order of business is to open a project. I want to give making those work items a try.

1:25pm. I am confused. I opened a new project, but for some reason the Team Explorer is connected to the old one. What the hell?

1:35pm. Well, nevermind that. I'll just use the existing bare bones project since that is what book is doing. I've figured out how to use the backlog as well.

1:45pm. I am really getting distracted by my own thoughs. Let me go to the next step.

2:20pm. 115/558. I am making slow progress, but my focus is starting to deepen. Yesterday was a snoozefest, and now I am finally starting to get into the mood again. There are a bunch of differences between the newest VS version that I have and the screenshots in the book, but it is quite similar apart from that.

2:30pm. "This meant that every application was tied to be hosted as a web application. ASP.NET Core 2.1 introduced the generic host, which allows for applications that do not necessarily have to process web-based HTTP requests"

This is something I will definitely have to study, and no doubt the book will cover it.

3:10pm. 119/558. I am making very slow progress through this book. But I do not find it tedious in the usual sense. Rather as I read this, I am thinking about Spiral, and how I would do this in that language.

```
inl (A:B:C:) = A + B + C
```

I think it really might be a good idea to allow this kind of syntax. It might improve the readability of the language.

I was looking forward to it back in v0.1 with objects, but it might be possible to use it with modules as well.

3:15pm. ASP.NET is all about DSL building from small pieces. I am thinking to the situation where I have large libraries whose functions have many arguments. I want to grab some benefits from Smalltalk.

Just like with union types, I can go further here. Without a doubt, I am going to make v0.2 the best version of Spiral yet.

3:35pm. 121/558. Do I have to necessarily use the MVC layout page, or will the Razor one do?

Also the screenshot shows the `wwwroot` folder, but I haven't received any instructions to actually create it. I guess I'll leave it for later.

3:45pm. 124/558. I am just following instructions without really understanding what they do, so this chapter is not too motivated, but I'll practice this craft. The closest I've come to think kind of programming are GUIs back in 2016.

3:50pm. There is no going through this quickly. I am in a pleasant mindset, so I'll go through this slow as I am now. My focus is not too high, and I am getting lost in daydreams, but I am not particularly bored.

4pm. 126/558. Let me take a break here. I think I will put in some overtime today since things are going to slowly, but aren't too tedious.""