"4:20pm. Let me resume.
4:25pm. 128/558. I put in npm file and the required packages. But I need to figure out what is going on with that `wwwroot` folder. Have I missed the instruction to create it somewhere?

4:30pm. Nope, there really is nothing about `wwwroot`. Let me create it then.

4:55pm. 130/558. So far I've been adding things by hand, but this is a bit over the line to input this way.

```
 <link rel="stylesheet" href="https://stackpath.bootstrapcdn
 .com/bootstrap/4.3.1/css/bootstrap.min.css"
 asp-fallback-href="~/lib/bootstrap/dist/css
 /bootstrap.min.css"
 asp-fallback-test-class="sr-only" asp-fallback-test-
 property="position" asp-fallback-test-value="absolute"
 crossorigin="anonymous"
 integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784
 /j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T"/></environment>
 <link rel="stylesheet" href="~/css/site.css" />

```

I think I'll just paste this.

5:15pm. This is not working for me. I like short instructions that are explained, but now I am just getting long dumps that are scarcely touched upon.

Right now I have an issue, I added the NPM packages before I created the `wwwroot/lib` folder and as a result they go put somewhere else. What do I do?

Well, what I will do is study the sample code.

5:20pm. After I am done with this chapter, I will try recreating what I've learned in F#.

The truth, I really do appreciate this sample application. As I was saying, I do not want to challenge the gambling dens right away, so I am open to training agents and renting them for a price. Everything here is a vital skill towards enabling that kind of business.

5:25pm. 135/558. Just how long till the end of this chapter. I want to skim this and then I'll take a look at the code samples.

On one hand, this project is simple. It is not like proving theorems in Coq. On the other hand, there are shitload of details to go through. It will take me a while to convert what I am doing now into actual skill.

5:55pm. 145/558. I am not really studying this intently anymore, I am just trying to get a lay of the land so to speak.

I admit, I didn't think this chapter would be so difficult considering the task is to make a Tic Tac Toe app. There is really a lot of stuff in it, and I may have to study for a few days before I move on. The reason for this is that a lot of the complexity seems to be front loaded.

Yesterday I was not sure whether I was making the right choice to switch from studying TS since the book said it required only a minimal familiarity with ASP.NET, but I definitely made the right choice. This basic familiarity is not as easy as it seems, and I definitely do need to take the effort to do this. Had I continued with TS, I'd definitely have run into difficulties so I am glad that I am finally doing things in proper order.

I am sure that after I really get this chapter down the following ones will be easier.

6:05pm. I've gotten lost in thought. Let me skim the last ten pages and I will call it a day. I'll leave the code examples and rebuilding it in F# tomorrow.

6:30pm. Sheesh. Finally done. I am tired at this point.

Maybe more than a few days, it will take me quite a bit of study until I understand enough to recreate the Tic Tac Toe without the help from the book."