"1:55pm. Done with breakfast, chores and everything else. Let me finish ep 8 of the Toji No Miko game and then I am going to resume my harvest of employable skills. Though I have a lot of programming related skills by now, this webdev stuff is important. If I tried looking for a job, my prospects would be a lot poorer without what I am going to learn in the next few months.
Well, it is just a contingency plan in case my computer dies or something like that.

2:15pm. That was fun. Let me finally start. Really, so much of learning is setting up the right curriculum for oneself. Everybody wants to be a genius and learn without effort. It is unacknowledged how much time gets wasted in the process of finding the right path.

Now where do I start?

With the ASP.NET documentation.

https://docs.microsoft.com/en-us/aspnet/#pivot=core&panel=core_tutorials

Since the book is too much, let me start with this.

https://docs.microsoft.com/hr-hr/aspnet/core/tutorials/razor-pages/razor-pages-start?view=aspnetcore-3.1&tabs=visual-studio

Let me go with this.

2:25pm. https://github.com/dotnet/AspNetCore.Docs/tree/master/aspnetcore/tutorials/razor-pages/razor-pages-start/sample/RazorPagesMovie30

Thankfully the docs have plenty of tutorial material.

2:35pm. I am just daydreaming here rather that studying. I'll get a move on.

Yeah, I can get through this with some handholding. The docs should be exactly what I need. First come the docs, then come the books. And I will want to take a look at the source AST.NET when all is done, just like for VS Code. I am going to aim for complete mastery here. Rather than just learning a trick or two, it would be far better to completely shift my mindset to the next level.

Back in 2016 I did GUIs and that poker game. Originally I wanted everything to be graphical, but ended up being glued to the command line. The Windows GUI stuff turns out to be extremely tedious - the effort it takes to get GUIs to be a benefit is completely out of proportion. Maybe it would be different if I was working in Pharo, but that platform is too messy for me to be serious about it.

Back then and over the years, I'd get a thought that rather than looking down on webdev so much, that I should consider the browser as my platform of choice. I'd watched the videos on the browsers of the future being complete OSes.

So maybe I should have Spiral v0.2 target webassembly assuming it has GC (I read that it does not.)

Regardless, maybe after I am done with this I might conclude that the web is the platform that I need for my experiments.

2:45pm. It is not like I am completely doing this on a whim. I let the waves guide me, but I am the one that decided whether to move forward or stop.

Let me finally try that Razor tutorial."