"2:50pm. Seriously, how do I disconnect from the Azure account? Every time I open a new project it remains connected to the old one.
...Uf I am not sure what to search for in Google. Well, forget that for now.

I'll deal with that later when I need to. For now focus on the tutorial.

3:10pm. Right now I am looking at various files including the bootstrap.js and jquery.js. Yeah, as bad as TS's OO stuff is, it is still a marked improvement over having a big library in JS with zero types.

```
<!DOCTYPE html>
<html lang="en">
<head>
    <meta charset="utf-8" />
    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
    <title>@ViewData["Title"] - RazorPagesMovie</title>
    <link rel="stylesheet" href="~/lib/bootstrap/dist/css/bootstrap.min.css" />
    <link rel="stylesheet" href="~/css/site.css" />
</head>
<body>
    <header>
        <nav class="navbar navbar-expand-sm navbar-toggleable-sm navbar-light bg-white border-bottom box-shadow mb-3">
            <div class="container">
                <a class="navbar-brand" asp-area="" asp-page="/Index">RazorPagesMovie</a>
                <button class="navbar-toggler" type="button" data-toggle="collapse" data-target=".navbar-collapse" aria-controls="navbarSupportedContent"
                        aria-expanded="false" aria-label="Toggle navigation">
                    <span class="navbar-toggler-icon"></span>
                </button>
                <div class="navbar-collapse collapse d-sm-inline-flex flex-sm-row-reverse">
                    <ul class="navbar-nav flex-grow-1">
                        <li class="nav-item">
                            <a class="nav-link text-dark" asp-area="" asp-page="/Index">Home</a>
                        </li>
                        <li class="nav-item">
                            <a class="nav-link text-dark" asp-area="" asp-page="/Privacy">Privacy</a>
                        </li>
                    </ul>
                </div>
            </div>
        </nav>
    </header>
    <div class="container">
        <main role="main" class="pb-3">
            @RenderBody()
        </main>
    </div>

    <footer class="border-top footer text-muted">
        <div class="container">
            &copy; 2020 - RazorPagesMovie - <a asp-area="" asp-page="/Privacy">Privacy</a>
        </div>
    </footer>

    <script src="~/lib/jquery/dist/jquery.min.js"></script>
    <script src="~/lib/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
    <script src="~/js/site.js" asp-append-version="true"></script>

    @RenderSection("Scripts", required: false)
</body>
</html>
```

This is actually readable unlike that the example from the book where it had machine generated content. Figuring this out should be doable.

I still have no idea what the bootstrap and jquery libraries are for. I'll get to that eventually.

3:15pm. Ok, ok, ok...

Stop reading random stuff. What does the tutorial say?

3:50pm. Had to take a break. Let me resume.

Today, my motivation and focus has been low. As expected, this is not my kind of programming and it will take a while till I get on track. But I will just keep plugging away at it.

Today, most of my effort has gone towards dreaming up what I will do with this webdev knowledge.

As long as proper ways of doing ML are still a mystery, I might as well focus on cleaning up all the auxiliaries. I'll do it. By the time 2022 arrives I will have something nice to show to the rest of the world.

4pm. https://docs.microsoft.com/hr-hr/aspnet/core/fundamentals/static-files?view=aspnetcore-3.1

Reading this stuff is beyond boring. Let me get back to the Razor tutorial. I'll figure the rest as I go along.

If it turns out that ASP.NET is not working for me, I'll ditch it for the time being and just study pure TS from one of the books.

At least with pure HTML and TS, I'll know where I stand and that will prepare me for this framework.

4:40pm. https://docs.microsoft.com/hr-hr/aspnet/core/tutorials/razor-pages/model?view=aspnetcore-3.1&tabs=visual-studio#test-the-app

I am getting a 404 error. This is not working for me.

My focus is dirt low.

I hate this.

I was prepared to learn HTML and TS/JS, but having these weird additions on the dotnet side, and having the IDE generate code is not my idea of programming. My immersion is horrible.

Ultimately, skill is something that is build from the ground up.

Forget learning complicated frameworks for the time being. I'll return to ASP.NET later. I can always look at the sample code for this to find out what is wrong.

I want and need to do this bottom up. Once you can trace what the program is doing every step of the way, then you can learn.

4:50pm. Incidentally, the inability to do this kind of learning is why I am so mad at modern math. Modern math is like these crappy framework tutorials. Do this, do that, magic happens and no whys are ever given.

The reason why Wildberger is a great lecturer is because he fundamentally understand what learning is about. He wants to integrate that understanding with what math itself should be. I am much the same. If I have a programming skill, I do want to generalize it. It is why I fell so hard for Spiral.

I am not just good at programming - I am programming. That is why I cannot abide by flaws in my technique. I am guessing math is the same for him.

4:55pm. Today seems like another botch. I should just move to a TS book, but tedium is really hitting me right now.

Once again, all this really shows how easy it is to go down the wrong path and cause time to be wasted. I really felt like I was getting flow back when I was reading the TS book. I need to get back to that.

5pm. Ouh...

But it really does seem like ASP.NET has a lot of what I would find useful. I need to level up first.

Uffff...this is the worst.

5:05pm. Let me do this - I'll give the first chapter of one of the TS books a try, so I can wisely spend whatever fuel I have left in me. "