"9:05am. Let me chill for a bit and then I will start. Yesterday went quite well. For once I am learning quite well, and internalizing the material solidly. Hopefully I've gone beyond wasting time looking at random things with this.
9:25am. Ok. Since I am in the mood I won't waste too much time. Let me start now. If I am lucky, I'll be able to go through entirety of part two in time for breakfast. Let me get this thing on.

10:15am. 186/551. Done with chapter 7. Let me continue onwards. There was a bunch new in chapter 7 that I did not see in either the TS docs or the previous books.

10:35am. 202/551. Done with chapter 8. This stuff is not too complicated. I've been through this before.

10:45am. Focus me, focus. Let me get through this chapter.

10:55am. 226/551. Done with chapter 9. Yeah, no way will I finish 200 pages in one morning. But 70 is not bad progress regardless. Let stop here for a while.

11:10am. Actually, let me go on for another chapter and then I will take that break. It is too early to stop right now.

11:35am. 253/551. Skimmed the previous chapter.

12pm. 283/551. The OO stuff is dull, but at least I got through it.

One thing I still do not get is why mutating the prototype of an object does not mutate the prototype itself. When making an object, does the Object prototype get copied.

Just like loop environments, JS is really wonky when it comes to scope. All the references are implicit and the code does not make logical sense as written.

Well, nevermind that. It is not like I am going to do significant programming in TS. I just have to understand it well enough to use it, I certainly won't ever need to mess with under the hood stuff.

12:20pm. 303/551.

```
filter<V extends T>(predicate: (target) => target is V): V[] {
    return this.items.filter(item => predicate(item)) as V[];
    }
```

Interesting bit of code. TS is quite powerful to be able to do this.

12:30pm. 312/551. Apart from that interesting tidbit, I admit I skipped most of the last chapter.

I am making really good progress through the book so far. All these advanced features do not really matter, as FP just much better and simpler solutions.

I am going to do this. I am going to master webdev. Lately my thought have been going back to 2016 and that poker game. It was really a pain in the ass to make that GUI for it. Even something as simple as charting the training progress makes .NET shit itself.

The web is much better suited for this kind of work.

This time I am going to do what I could not do back then, and in Pharo.

I am going to do every single thing right. And I am going to draw out the true power of those neurochips.

I've been thinking about getting a job at Intel's neurochip division, but on second thought that sort of thing does not suit my character. Rather than mess with that, I should just inquire whether they would be open to sponsoring my work. I really am a PL researcher at this point, and before I can move on from this phase of my life, I really have to finish it all.

Having a boss telling me what to do is not a role I should be aspiring to. Paid research is a thing in this world.

Before I do that though, I actually want to finish v0.2. I want something that I can show to other people. So my pitch will be a mix of the new, and samples from my old work. I think it should be decently impressive when I show them those Cuda kernels doing AD without a single type annotation in sight.

What I will be applying to is to do what I did in 2018 in Cuda for GPU, to do in their language for neurochips. I think the pitch is good, but it is not a guaranteed success. For all I know, the guys making the hiring decisions might not see the importance of this, or look down on programming. Maybe they'll want to monkey it all the way through. Who knows.

All I really need is that next gen piece of hardware and I will be set. It does not have to be Intel, surely I'll hit 1 out of 100.

Unlike the rest of the ML crowd, I will easily be able to write my own libraries when I have Spiral v0.2 ready. I do not need a framework for this."