"8:30am. I am up quite early today despite going to bed at 12pm yesterday. I have Cultivation Chat Group a try yesterday and went on a binge. It was quite nice. I am not a fan of cultivation stuff, but the urban setting and the novelty of the genre itself was enough to hold my interest.
https://mangadex.org/title/44858
Blattodea

Finally this is getting started. And a new Toji OVA got announced.

8:40am. Ok, today I will chill for a bit. I am not doing enough chilling.

10:15am. I've done enough chilling.

https://www.youtube.com/watch?v=otoCxbZel1o&feature=youtu.be
Building High performance AI Training Systems with Gaudi

But rather than webdev, let me watch this. A recent ML sub thread made me want to look up Habana.

10:35am. Ok, unlike Mike Davies talk, the 10m of this is enough to know how all of it will go. The product presented here is really quite straightforward and not really as interesting as SNN processors. More speculative devices like Loihi are definitely more to my liking.

Interesting that they have much better latency, but it is not all the way for a batch size of 1. If I had to use something for RL, I'd pick this, but it is not a such a massive leap forward compared to GPUs like what Davies showed with SNNs.

10:40am. Let me put in an hour into studying the webdev chapter. It is finally time to get this on the way."