"11:10am. 381/551. My focus is not that high here, but I am not completely confused here unlike with the ASP.NET tutorials.
Though OO, this kind of code has a link to the root that I can follow. I can imagine myself writing something of this level of complexity.

11:25am. I am distracted with thoughts of Spiral.

I am thinking about functions, more specifically the way they can be returned from join points, if statements and put into arrays.

The fact this, though this use is allowed, there is not much utility to it. In fact, many times returning functions from join points happens by accident and it is usually a mistake.

...Hmmm...I remember. I think that in v0.09 I wanted to preserve the ability to put things like tensors into arrays.

11:25am. In that case why not `dyn` it and turn it into a reified join point?

That would make sure that I am not making any mistakes.

Ah, no wait. Right now the dyn behavior does memoization.

Drat. Really, functions are so troublesome.

I am going to restrict their use. In v0.2 I do not need to be as permissive as I was before.

11:30am. Yeah, in v0.09 I wanted to do as much experimentation as possible, but in v0.2 I am going to tighten the leash on the language. The focus is not power anymore, but ease of use. I need to be truthful to my design priorities.

11:35am. Let me add this TODO. That will put my heart at ease. I'll go with the assumption that functions will be restricted from here on out."