"9:45am. Today I got up late it seems. Let me take the time to make that reply to the PM and then I will start with the book.
```
> Have you found some areas where inlining and join wouldn't be automatically infereable (or at least very hard)?

> It seems to me that join points are trivial to locate.

The easy way of doing it would be to just put them at the end of every `let` function statement and have the optimizer decide when to inline them. The point of Spiral though is that it allows the programmer to decide where to put them.

Also the difficulty in join points is not just where to put them, but whether to convert functions to heap allcoated closures, and whether to specialize to constants or push them to runtime. This is a lot harder problem. In the case of closure conversion, the v0.09 partial evaluator does not even have the information necessary to do this and so has to be done by hand.

> Is inlining powerful enough to replace any syntactic macros?

No, but conversely to get the same power that Spiral's partial evaluation offers, you'd actually have to reimplement those features via macros. Macros do not memoize the results of their evaluation - this is what join points actually are. Make macros lexically scoped (hygienic) and allow them to do that memoization, and you have the core of Spiral's functionality.

> What is the difference between specialization and inlining?

In the v0.09 documentation, I kind of mushed the meaning of the two together even when they aren't. Back when I was writing it, I was putting in the effort to make up my own explanation for what Spiral does because no other language has its feature set, but let me see if I can get my vocabulary in order.

Inlining is just when you inline the body of the function, and specialization is what happens in join points. Also it is what happens when you instantiate reified join points (layout types in v0.09).

```
inl add a b = a + b
inl _ = add 1 2, add 1.1 2.2 // inlining - the bodies will get inlined
```

```
inl add a b = join a + b
inl _ = add 1 2, add 1.1 2.2 // specialization - join points will get compiled to two function calls during code gen
```

> Did you try to replace type inference with partial evaluation?

Yeah, in v0.09 type inference is done via partial evaluation. In effect, the way it happens is similar to in dynamic languages in that you have to run the program to get the type errors, except it happens at compile time. I call it bottom-up type inference as opposed to top-down.

After programming like that for a good while back in 2018, I've concluded that it is too hard after all. One of the difficulties is that unlike in dynamic languages, you can't just put anything into an array, you have to specify its type somehow. So when doing generic functions (think of `Array.map`) that means having to run them at the type level and then once you have the array type at the value level. This is just the start of it, I had to do some gross hacks to infer the types for some functions. If you look at v0.09 you will see that it has `type_raise` and `type_catch` which allow me to do type inference by throwing exceptions. And I did even worse things in some cases.

It was interesting at the time, but good programming should be easy. Looking back, it is obvious to me now that I feel in love with experimentation a bit too much.

> Can you do runtime specialization?

No. It would be harder and not serve the same purpose. There are systems like Lancet which can, but in Spiral I only wanted to build a bridge from high level F# code to Cuda GPU kernels in order to make a ML library that does not rely on frameworks. This was my main goal, and I did not want to distract myself my making what is essentially a JIT compiler.

> How do you deallocate memory?

On the F# side, the GC takes care of that. In the ML library you do have to do it yourself for GPU memory, but I consider that a mistake. I should have used weak arrays in the GPU allocator and had the .NET GC take care of that as well. But I wasn't sure whether that sort of scheme would be reliable enough. Python is really lucky that it does reference counting and has reliable GC as a result.

I am going to have to do more research here - I definitely want full GC for my next ML library. Maybe it won't be needed though. For the sake of performance, in the next gen hardware effort is being made to localize memory to the cores rather than having them allocate from global pool. Meaning [neurochips](https://www.youtube.com/watch?v=jhQgElvtb1s) and [AI accelerators](https://www.youtube.com/watch?v=otoCxbZel1o) will (in programming terms) have static allocation (similar to shared and local Cuda memory).

Spiral is really well suited for this kind of thing by the way - if you look at how the ML code gets compiled you will see that all the tensors have their dimensions specialized in the GPU kernels. In some cases this is actually necessary in order to properly allocate shared memory. It also leads to less register use.

> How do you mutate memory?

Similarly to how the MLs and F# does it. References and arrays can be mutated directly. In GPU kernels the same philosophy is followed except they are statically allocated in registers.

There are also records with mutable fields, but I need to clean those up for v0.2.

I really like this explicit reference approach to mutation pioneered by the ML family of languages. Right now I am learning Typescript and everything is a huge mess as all the references are implicit.

I don't like how Haskell does it either as it is too rigid. Monadic code leads to coupling which is bad. I really made the right choice to write Spiral in F# as opposed to Haskell.
```

10:50am. The above is what I had been writing for the past hour. At my discretion, I'll paste it here in this journal since none of the questions are sensitive. If they were I'd have kept it private.

10:55am. Ok...I am thinking of whether to take a break here.

How about I give the Angular chapter a try for an hour and then I'll do that. I want to get this started.

11:45am. I've skimmed the rest of the book.

I really thought I would get a lot of out of the framework parts, but as it turns out I do want to leave this for when I actually need it.

With this I am finally done with the Essential TS book. Great.

Next up: the basics of web development. I am going to build my skills bottom up.

https://svelte.dev/

If I had to use a framework, I'd rather go for something which is a language. So maybe something like Elm or this.

11:55am. Let me take that break. I think I'll do the review today. Given that reply, and finishing the book ahead of schedule, today seems the ideal time to do it before I start on the next part. After that, I'll go through the video and go through those links.

It should not me too long to figure out the basics.

But ASP.NET after that should be more challenging. Still, I am looking forward to mastering this.

Just this thing and I will be able to do anything."