"1:20pm. Let me do the chores.
1:45pm. Finally done. Let me chill just a little bit and then I am going to start.

...Let me do it.

Time to clear that last DOM video. As tedious as this is, I need to continue studying.

2:05pm. Ok, rather than falling asleep here, how about I follow along instead?

I am really impressed what a difference importing that bootstrap template makes.

2:30pm. Done with the video. Am slacking a bit right now, but I'll start the next lecture soon.

2:45pm. https://www.youtube.com/watch?v=qi9VQqYcXqY
What To Learn Before A JavaScript Framework

Let me watch this and then I'll pick a framework. I want just one, not all 3 of them. After that I'll go for Nodejs. And then comes ASP.NET. Hopefully now that I've gone through this tutorial round, I will have an easier time knowing what things are.

3:15pm. Had to take a break. Let me finish the video and then I'll finally start.

3:25pm. Done with the vid. Let me go through the Vue crash course. It is shorter than the other two, was mentioned to be easier to learn, and I do not particularly care about any framework in particular.

It is only a bit longer than an hour so I should be able to deal with it today. This will leave me free to deal with Nodejs tomorrow. After that I can move to ASP.NET. Though before that I do want to take a look at the MDN documentation, just to shore up my knowledge of HTML and CSS a bit more. Still, by now I know what this is all about.

3:35pm. https://youtu.be/Wy9q22isx3U?t=311

This HTML interpolation, is that a part of the actual HTML standard?

https://youtu.be/Wy9q22isx3U?t=346

The list here is interesting. There is even a GUI app for making interfaces?

3:50pm. https://youtu.be/Wy9q22isx3U?t=781

This is actually pretty impressive. Quite a bit more than I expected.

Now since I finally got the stuff installed, let me backtrack a bit so I can follow the UI steps from the start.

3:55pm. It sure takes it a while to setup a project.

4pm. And right away I run into some bug. For some reason it does not move to the next step even though I've created the project.

4:20pm.

```
<template>
  <div class="hello">
    <h1>{{ msg }}</h1>
    <p>
      For a guide and recipes on how to configure / customize this project,<br>
      check out the
      <a href="https://cli.vuejs.org" target="_blank" rel="noopener">vue-cli documentation</a>.
    </p>
    <h3>Installed CLI Plugins</h3>
    <ul>
      <li><a href="https://github.com/vuejs/vue-cli/tree/dev/packages/%40vue/cli-plugin-babel" target="_blank" rel="noopener">babel</a></li>
      <li><a href="https://github.com/vuejs/vue-cli/tree/dev/packages/%40vue/cli-plugin-eslint" target="_blank" rel="noopener">eslint</a></li>
    </ul>
    <h3>Essential Links</h3>
    <ul>
      <li><a href="https://vuejs.org" target="_blank" rel="noopener">Core Docs</a></li>
      <li><a href="https://forum.vuejs.org" target="_blank" rel="noopener">Forum</a></li>
      <li><a href="https://chat.vuejs.org" target="_blank" rel="noopener">Community Chat</a></li>
      <li><a href="https://twitter.com/vuejs" target="_blank" rel="noopener">Twitter</a></li>
      <li><a href="https://news.vuejs.org" target="_blank" rel="noopener">News</a></li>
    </ul>
    <h3>Ecosystem</h3>
    <ul>
      <li><a href="https://router.vuejs.org" target="_blank" rel="noopener">vue-router</a></li>
      <li><a href="https://vuex.vuejs.org" target="_blank" rel="noopener">vuex</a></li>
      <li><a href="https://github.com/vuejs/vue-devtools#vue-devtools" target="_blank" rel="noopener">vue-devtools</a></li>
      <li><a href="https://vue-loader.vuejs.org" target="_blank" rel="noopener">vue-loader</a></li>
      <li><a href="https://github.com/vuejs/awesome-vue" target="_blank" rel="noopener">awesome-vue</a></li>
    </ul>
  </div>
</template>

<script>
export default {
  name: 'HelloWorld',
  props: {
    msg: String
  }
}
</script>

<!-- Add "scoped" attribute to limit CSS to this component only -->
<style scoped>
h3 {
  margin: 40px 0 0;
}
ul {
  list-style-type: none;
  padding: 0;
}
li {
  display: inline-block;
  margin: 0 10px;
}
a {
  color: #42b983;
}
</style>
```

This is actually pretty cool. I can see how this would make programming easier. I approve. This is definitely worth using over vanilla HTML+CSS+JS.

But given how the framework works, I really would say that this is a different language rather than a straightforward extension. But ASP.NET is like that as well, and everything of value could be expected to be like this.

Also, I was wondering about HTML interpolation, but even though I had not looked it up, I am 95% sure that some of the stuff here is Vue specific.

Actually, I've really been yearning something like this back when I was doing desktop GUIs. For this kind of tooling. For this kind of experience.

4:50pm.

```
  components: {
    Todos
  },
```

What is this syntax? Is it something like `Todos: Todos`?

5pm.

```
        <div v-for="todo in todos" v-bind:key="todo">
            <h3>{{todo.title}} </h3>
        </div>
```

I can't get it to work the same as in the video. But my version is a lot newer so maybe it is because of that.

5:05pm.

```
<template>
  <div id="app">
    <Todos v-bind:todos="todos"/>
  </div>
</template>

<script>
import Todos from './components/Todos'

export default {
  name: 'App',
  components: {
    Todos
  },
```

```
<template>
    <div>
        <div v-for="todo in todos" v-bind:key="todo">
            <h3>{{todo.title}} </h3>
        </div>
    </div>
</template>

<script>
export default {
    name: "Todos",
    props: ["todos"]
}
</script>
```

There is a lot of circularity going on around here. Does the same thing have to be bound and referenced in so many different places?

5:10pm. The Vue plugin just went bonkers.

5:15pm. I can't figure out what is wrong with it. At any rate it does not matter. Let me just go with the video. I'll ignore the errors.

5:20pm. Ah, it makes a difference whether the file extension is `.Vue` vs `.vue`. I did the first one by accident.

5:50pm. https://youtu.be/Wy9q22isx3U?t=2390

I am making slow progress, but I am totally into this. This is quite immersive. I can really see the way this framework flows, but I am wondering whether this is the most elegant solution.

Since this is this interesting, this is making me motivated to check out the other frameworks.

...I am thinking whether I should stop here. Let me have lunch and then I'll resume. I'll put in some overtime today.

6:15pm. Let me resume. I can still go on for a while longer.

7:10pm. https://youtu.be/Wy9q22isx3U?t=3588

Only 9m left. I really am putting in overtime today. But this framework stuff really comes live to me when I am watchin a video as opposed to following a book.

I am not really satisfied with Vue's elegance. There is a bit too much bungling with variable bindings. I am going to repeat this exercise with Svelte and the other two frameworks, though I am not expecting Angular and Reach to be better.

7:30pm. Finally done. I thought it would never end.

I wasn't even slacking, but it took me like 3h to go through a single video. Honestly, I feel that there is room for improvement here, though Vue is definitely seems better than the raw experience.

I am going to just take a look at the Reach and Angular videos for half an each tomorrow, and then take a look at Svelte since he mentioned it. I much rather prefer compilers rather than frameworks and would not mind a step up from this.

Is the monthly PL thread up yet? Nope. I'll leave the posting the review for tomorrow then.

7:35pm. With this I am done. Today went well. The DOM videos put me to sleep, but even with them I learned a few new tricks.

I am getting closer to my goal of studying the backend stuff."