"8:35am. I am up. Yesterday was a mess, but there is no doubt about it - the UI difficulties that have plagued me for years are going to be resolved this month. Without a doubt.
The Svelte guy deserves great praise for his accomplishment.

If I weren't obsessed about drawing out the power of the next gen hardware for the sake of AI, and was a webdev instead, I'd like to think that I would have made an UI language much like Svelte.

I mean, what he did might seem simple, but webdev has existed for decades and yet only now this comes out.

Being able to do things in less than 100 lines of code that would have previously taken 1000 is a game changer for me.

Blazor has a similar design, so if I want to stay with .NET, I can just go with it.

8:40am. It is really wonderful, I have two, not just one options that are both decent when it comes to UI creation.

I said that Pharo has good UI support, but not like this. This - this is what I wanted all along.

Before this, the only real choice in my mind was Elm, but having a virtual DOM and diffing on every pass while elegant and effective is only a 2/3rds of a solution at most. It is a solution, but it does not quite capture what UI programming really is. It more like works around the issues.

The Svelte design is without a doubt 100% there. It is perfect.

8:45am. I think I can allow myself to get excited about programming once more. Give me a year and I will have complete mastery of web development.

Without a doubt, I will try ML once again and this time I am not going to end up being glued to the command line.

Before I start chilling, let me say one thing - I am surprised how similar desktop UI development and web development are. I thought there would be more pronounced differences, but I was wrong. Desktop might lack HTML and CSS, but besides that the programming model is identical.

I guess in the end, me looking down on webdev was merely ignorance.

8:50am. Now, let me chill for a bit and I will start. I am actually motivated today, so I won't dawdle too much."